IMP: draw textures of all itens
