+++ resources/textures/elements/espada.png
