+++ resources/textures/monster/monstro-norte.png
