+++ resources/textures/monster/monstro-oeste.png
