+++ resources/textures/monster/monstro-sul.png
