+++ resources/textures/player/jogador-leste.png
