+++ resources/textures/player/jogador-norte.png
