+++ resources/textures/player/jogador-oeste.png
