+++ resources/textures/player/jogador-sul.png
